Reformatted node core and exported to another file
