+++ Artificial_skin_PC_visualisation/nodes/node_core.py
@@ -1,0 +1,102 @@
+from enum import Enum
+from PyQt5 import QtCore
+
+import rospy
+
+from debug.debug import *
+
+
+class Topic:
+    def __init__(self, name, msg_type, callback=None, queue_size=10):
+        self.name = name
+        self.msg_type = msg_type
+        self.callback = callback
+        self.queue_size = queue_size
+
+
+class NodeStatus(Enum):
+    unknown = 0
+    not_connected = 1
+    connected = 2
+    connection_crashed = 3
+    connection_bad_messages = 4
+    working = 10
+    calibrating = 11
+
+
+class Node(QtCore.QThread):
+    language = "en"
+    translation = None
+
+    subscribed_topics = []
+    published_topics = []
+    subscribers = []
+    publishers = []
+
+    def __init__(self, node_name, subscribed_topics=None, published_topics=None, language="en"):
+        super(Node, self).__init__()
+        rospy.init_node(node_name)
+
+        # Place where you should import all translations
+        self.language = language
+        if self.language == "en":
+            from languages import en as translation
+            self.translation = translation
+        else:
+            from languages import en as translation
+            self.translation = translation
+
+        # Setup subscribed topics
+        if subscribed_topics is not None:
+            self.subscribed_topics = subscribed_topics
+
+        # Setup published topics
+        if published_topics is not None:
+            self.published_topics = published_topics
+
+        # Initialize all topics
+        for topic in self.subscribed_topics:
+            sub = self.Subscriber(topic)
+            self.subscribers.append(sub)
+        for topic in self.published_topics:
+            pub = self.Publisher(topic)
+            self.publishers.append(pub)
+
+        self.exitFlag = False
+        self.start()
+
+    def publish_msg_on_topic(self, topic_name, msg):
+        for pub in self.publishers:
+            if pub.topic.name == topic_name:
+                pub.publish(msg)
+
+    def get_node_status(self):
+        # TODO describe and use this statuses in code
+        return self.translation.nodeStatusTranslationDictionary[self.node_status]
+
+    class Subscriber:
+        topic = None
+        callback_function = None
+
+        def __init__(self, topic):
+            self.topic = topic
+            self.callback_function = topic.callback
+            self.sub = rospy.Subscriber(self.topic.name, self.topic.msg_type, self.subscriber_callback)
+
+        def subscriber_callback(self, data):
+            self.callback_function(data)
+
+    class Publisher:
+        topic = None
+
+        def __init__(self, topic):
+            self.topic = topic
+            self.pub = rospy.Publisher(topic.name, topic.msg_type, queue_size=topic.queue_size)
+
+        def publish(self, message):
+            if type(message) is self.topic.msg_type:
+                self.pub.publish(message)
+            else:
+                debug(DBGLevel.WARN,
+                      "Invalid publish message type, expected: " + str(self.topic.msg_type) + ", received: " + str(
+                          message) + ".")
